Few possibly useful improvements
shortix.sh:
- remember the last update and only run if compatdata has newer directories
- don't break with read when protontricks could not be found (service?)
- check for existence of compatdata

shortx_install.sh
- fixed systems ->systemd typo
- ask if the service should be set up...

+++ shortix_install.sh
@@ -1,4 +1,5 @@
 #/bin/bash
+
 if [ -d $HOME/Shortix/ ] || [ -f $HOME/.config/systemd/user/shortix.service ]; then
   echo "########################"
   echo "#   UPDATING SHORTIX   #"
@@ -13,14 +14,20 @@ fi
 mkdir -p $HOME/Shortix
 cp /tmp/shortix/shortix.sh $HOME/Shortix
 chmod +x $HOME/Shortix/shortix.sh
-if [ ! -d $HOME/.config/systemd/user ]; then
-    mkdir -p $HOME/.config/systems/user
+
+read -e -n 1 -p "Would you like to setup system service for background updates [Y/n]: "
+if [[ ! $REPLY =~ ^[Nn]$ ]]; then
+  if [ ! -d $HOME/.config/systemd/user ]; then
+      mkdir -p $HOME/.config/systemd/user
+  fi
+  cp /tmp/shortix/shortix.service $HOME/.config/systemd/user
+  systemctl --user daemon-reload
+  if [ ! systemctl is-enabled --user  shortix.service ]; then
+    systemctl --user enable shortix.service
+  fi
+  systemctl --user restart shortix.service
 fi
-cp /tmp/shortix/shortix.service $HOME/.config/systemd/user
-systemctl --user daemon-reload
-if [ ! systemctl is-enabled --user  shortix.service ]; then
-  systemctl --user enable shortix.service
-fi
+
 if [ -f $HOME/.config/user-dirs.dirs ]; then
   source $HOME/.config/user-dirs.dirs
   if [ $XDG_DESKTOP_DIR/shortix_installer.desktop ]; then
@@ -30,5 +37,5 @@ if [ -f $HOME/.config/user-dirs.dirs ]; then
     chmod +x $XDG_DESKTOP_DIR/shortix_updater.desktop
   fi
 fi
-systemctl --user restart shortix.service
+
 echo  "DONE!"
